Effects.System.PosixCompat

Description

Provides the MonadPosixCompat typeclass.

Since: 0.1

Effect

class Monad m => MonadPosixCompat m where

Class for unix-compat effects.

Since: 0.1

Methods

setFileMode :: HasCallStack => FilePath -> FileMode -> m ()

Since: 0.1

setFdMode :: HasCallStack => Fd -> FileMode -> m ()

Since: 0.1

setFileCreationMask :: HasCallStack => FileMode -> m FileMode

Since: 0.1

fileAccess :: HasCallStack => FilePath -> Bool -> Bool -> Bool -> m Bool

Since: 0.1

fileExist :: HasCallStack => FilePath -> m Bool

Since: 0.1

getFileStatus :: HasCallStack => FilePath -> m FileStatus

Since: 0.1

getFdStatus :: HasCallStack => Fd -> m FileStatus

Since: 0.1

getSymbolicLinkStatus :: HasCallStack => FilePath -> m FileStatus

Since: 0.1

createNamedPipe :: HasCallStack => FilePath -> FileMode -> m ()

Since: 0.1

createDevice :: HasCallStack => FilePath -> FileMode -> DeviceID -> m ()

Since: 0.1

createLink :: HasCallStack => FilePath -> FilePath -> m ()

Since: 0.1

removeLink :: HasCallStack => FilePath -> m ()

Since: 0.1

createSymbolicLink :: HasCallStack => FilePath -> FilePath -> m ()

Since: 0.1

readSymbolicLink :: HasCallStack => FilePath -> m FilePath

Since: 0.1

rename :: HasCallStack => FilePath -> FilePath -> m ()

Since: 0.1

setOwnerAndGroup :: HasCallStack => FilePath -> UserID -> GroupID -> m ()

Since: 0.1

setFdOwnerAndGroup :: HasCallStack => Fd -> UserID -> GroupID -> m ()

Since: 0.1

setSymbolicLinkOwnerAndGroup :: HasCallStack => FilePath -> UserID -> GroupID -> m ()

Since: 0.1

setFileTimes :: HasCallStack => FilePath -> EpochTime -> EpochTime -> m ()

Since: 0.1

touchFile :: HasCallStack => FilePath -> m ()

Since: 0.1

setFileSize :: HasCallStack => FilePath -> FileOffset -> m ()

Since: 0.1

setFdSize :: HasCallStack => Fd -> FileOffset -> m ()

Since: 0.1

getPathVar :: HasCallStack => FilePath -> PathVar -> m Limit

Since: 0.1

getFdPathVar :: HasCallStack => Fd -> PathVar -> m Limit

Since: 0.1

PathType

data PathType

Path type.

Since: 0.1

Constructors

PathTypeFile
Since: 0.1

PathTypeDirectory
Since: 0.1

PathTypeSymbolicLink
Since: 0.1

PathTypeOther
Since: 0.1

displayPathType :: IsString a => PathType -> a

String representation of PathType.

Since: 0.1

Functions

throwIfWrongPathType :: (HasCallStack, MonadCatch m, MonadPosixCompat m) => String -> PathType -> FilePath -> m ()

Throws IOException if the path does not exist or the expected path type does not match actual.

Since: 0.1

isPathType :: (HasCallStack, MonadPosixCompat m) => PathType -> FilePath -> m Bool

Checks that the path type matches the expectation. Throws IOException if the path does not exist or the type cannot be detected.

Since: 0.1

getPathType :: (HasCallStack, MonadPosixCompat m) => FilePath -> m PathType

Returns the type for a given path without following symlinks. Throws IOException if the path does not exist or the type cannot be detected.

Since: 0.1

Optics

_PathTypeFile :: Prism' PathType ()

Since: 0.1

_PathTypeDirectory :: Prism' PathType ()

Since: 0.1

_PathTypeSymbolicLink :: Prism' PathType ()

Since: 0.1

_PathTypeOther :: Prism' PathType ()

Since: 0.1

Utils

throwPathIOError

Arguments

:: (HasCallStack, MonadThrow m)

=> FilePath
Path upon which the IO operation failed.

-> String
String location (e.g. function name).

-> IOErrorType
Type of exception.

-> String
Description.

-> m a

Helper for throwing IOException.

Since: 0.1